numberPerDosage in DrugOrders in EncounterTransactions should be a Double, not an Integer - EA-18

+++ api/src/test/java/org/openmrs/module/emrapi/encounter/domain/DrugOrderBuilder.java
@@ -13,7 +13,7 @@ public class DrugOrderBuilder {
     }
 
     public DrugOrderBuilder withBasicValues(String uuid, String conceptUuId, Date startDate, Date endDate,
-                                            Integer numberPerDosage,
+                                            Double numberPerDosage,
                                             String dosageInstructionUuid,
                                             String dosageFrequencyUuId){
         this.drugOrder.setUuid(uuid);
